Backend/requirements.txt
Django==4.2
asgiref
distlib
filelock
platformdirs
sqlparse
tzdata
virtualenv
pillow
django-simple-captcha
django-simple-math-captcha
django-robots
djangorestframework
djangorestframework-simplejwt
markdown
django-filter
drf-yasg[validation]
django-mail-templated
django-extensions
django-multi-captcha-admin
black
flake8
openpyxl
django-cors-headers
pandas
